// src/front/src/api/answerrule/models/dayOfWeek.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PBX AnswerRule API (via Gateway)
 * OpenAPI spec version: 1.0
 */

export type DayOfWeek = (typeof DayOfWeek)[keyof typeof DayOfWeek];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DayOfWeek = {
	Sunday: "Sunday",
	Monday: "Monday",
	Tuesday: "Tuesday",
	Wednesday: "Wednesday",
	Thursday: "Thursday",
	Friday: "Friday",
	Saturday: "Saturday",
} as const;